requirements.txt
pyyaml
spotipy
requests
